* Refactored tests based on IQ data to lighten the code.
